+++ packages/terra-date-input/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Changed
+* Update docs to use MDX syntax
 
 1.9.0 - (January 28, 2020)
 ------------------