+++ packages/terra-date-picker/CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 ### Changed
 * Changed `onChange` and `onChangeRaw` callbacks to return the following metadata as part of the third parameter: iso, inputValue, isCompleteValue, isValidValue
+* Update docs to use mdx syntax.
 
 4.27.0 - (February 4, 2020)
 ------------------